+++ classes/class.ilPanoptoPlugin.php
@@ -42,6 +42,7 @@ class ilPanoptoPlugin extends ilRepositoryObjectPlugin {
         global $DIC;
         $DIC->database()->dropTable(xpanConfig::DB_TABLE_NAME);
         $DIC->database()->dropTable(xpanSettings::DB_TABLE_NAME);
+        $DIC->database()->dropTable(SorterEntry::TABLE_NAME);
     }